add icons to menu left
